requirements.txt
kagglehub
pandas
mysql-connector-python
ansible